requirements.txt
Flask==0.12.0
Flask-Cache==0.13.1
CTRegisterMicroserviceFlask==0.3.0
requests==2.18
httmock==1.2.6
python-dateutil==2.6.1
mercantile==0.10.0
pyproj==1.9.5.1
opencv-python==3.4.0.12
colour==0.1.1
redis==2.10.6
google-cloud==0.27
urllib3==1.21.1
numpy==1.14.0rc1
matplotlib==2.1.1
